requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu117

# Core dependencies (minimal)
torch==1.13.1+cu117
torchvision==0.14.1+cu117
pytorch-lightning<2.0.0
opencv-python==4.9.0.80
einops==0.8.0
matplotlib==3.5.3
efficientnet_pytorch==0.7.1

# LiDAR or 3D point clouds
open3d==0.13.0
spconv-cu117
pyquaternion==0.9.9

# Geospatial (GPS-based) data
shapely==2.0.7
geopandas==0.13.2
haversine==2.8.1

# Hydra
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

# Utilities
fire==0.6.0
tensorboard
tensorboardX
pyrootutils
pre-commit
click
rich

# Local library (relative path)
-e ./fisheye_tools